refactor(echo): colocate atom internals under entrypoint modules.
Move Obj/Ref/Query atom helpers into internal/Obj, internal/Ref, and internal/Query so capitalized internal modules align with their public entrypoints.
