+++ packages/core/echo/echo/src/internal/Query/atoms.ts
@@ -7,12 +7,12 @@ import * as Atom from '@effect-atom/atom/Atom';
 import { URI } from '@dxos/keys';
 import { WeakDictionary } from '@dxos/util';
 
-import * as Database from '../Database';
-import type * as Entity from '../Entity';
-import type * as Filter from '../Filter';
-import * as Query from '../Query';
-import type * as QueryResult from '../QueryResult';
-import * as Registry from '../Registry';
+import * as Database from '../../Database';
+import type * as Entity from '../../Entity';
+import type * as Filter from '../../Filter';
+import * as Query from '../../Query';
+import type * as QueryResult from '../../QueryResult';
+import * as Registry from '../../Registry';
 
 // Keyed by queryable identifier. Holds the Queryable weakly so it is collected with its space.
 const queryableRegistry = new WeakDictionary<string, Database.Queryable>();
