+++ packages/core/echo/echo/src/internal/Query/index.ts
@@ -1,0 +1,5 @@
+//
+// Copyright 2024 DXOS.org
+//
+
+export * from './pretty';
