+++ packages/core/echo/echo/src/internal/Query/pretty.ts
@@ -1,5 +1,5 @@
 //
-// Copyright 2025 DXOS.org
+// Copyright 2024 DXOS.org
 //
 
 import { type QueryAST } from '@dxos/echo-protocol';
