+++ packages/core/echo/echo/src/internal/Ref/utils.ts
@@ -4,7 +4,7 @@
 
 import type * as Atom from '@effect-atom/atom/Atom';
 
-import type { Ref } from './Ref/ref';
+import type { Ref } from './ref';
 
 /**
  * Internal helper for loading ref targets in atoms.
